dont remove docker-production-aws folder

+++ docker-production-aws/microtrader-setup.sh
@@ -2,10 +2,13 @@
 
 # microtrader-setup.sh in https://github.com/wilsonmar/DevSecOps/docker-production-aws
 # as described in https://wilsonmar.github.io/docker-production-aws.
-# This script automates Justin Menga's course released 10 May 2016 on Pluralsight at:
+
+# This script builds from source the microtrader sample app used in 
+# Justin Menga's course released 10 May 2016 on Pluralsight at:
 # https://app.pluralsight.com/library/courses/docker-production-using-amazon-web-services/table-of-contents
 # which shows how to install a "microtraders" Java app in aws using Docker, ECS, CloudFormation, etc.
 
+# This script is run by this command on MacOS/Linux Terminal:
 # sh -c "$(curl -fsSL https://raw.githubusercontent.com/wilsonmar/DevSecOps/master/docker-production-aws/docker-production-aws-setup.sh)"
 
 # This is free software; see the source for copying conditions. There is NO
@@ -13,16 +16,12 @@
 
 # CURRENT STATUS: under construction, with TODO items.
 
-### 1. Parameters controlling this run:
+### 1. Run Parameters controlling this run:
 
 INSTALL_UTILITIES="no"  # or yes
-WORK_FOLDER="docker-production-aws"
+WORK_FOLDER="docker-production-aws"  # as specified in course materials.
 
-# TODO: MacOS version:
-
-### 2. Utilities:
-
-   ######### Starting time stamp, OS versions, command attributes:
+### 2. Context: Starting time stamp, OS versions, command attributes:
 
 clear  # screen
 # For Git on Windows, see http://www.rolandfg.net/2014/05/04/intellij-idea-and-git-on-windows/
@@ -37,22 +36,69 @@ SSH_USER="$USER@$( uname -n )"  # computer node name.
 echo ">>> $LOGFILE"
 #echo "$TO_PRINT" >$LOGFILE  # single > for new file
 echo "$TO_PRINT"  # to screen
+uname -a  # operating sytem
+
+
+### 3. Shell utility functions:
+
+cleanup() {
+    err=$?
+    fancy_echo "At cleanup() LOGFILE=$LOGFILE"
+    #open -a "Atom" $LOGFILE
+    open -e $LOGFILE  # open for edit using TextEdit
+    #rm $LOGFILE
+   FREE_DISKBLOCKS_END=$( df | awk '{print $4}' | cut -d' ' -f 6 ) 
+#   DIFF=$(((FREE_DISKBLOCKS_START-FREE_DISKBLOCKS_END)/2048))
+#   fancy_echo "$DIFF MB of disk space consumed during this script run." >>$LOGFILE
+      # 380691344 / 182G = 2091710.681318681318681 blocks per GB
+      # 182*1024=186368 MB
+      # 380691344 / 186368 G = 2042 blocks per MB
+
+   TIME_END=$( date -u +%s );
+   DIFF=$(( TIME_END - TIME_START ))
+   MSG="End of script $THISPGM after $(( DIFF/60 ))m $(( DIFF%60 ))s seconds elapsed."
+   fancy_echo "$MSG"
+   echo -e "\n$MSG" >>$LOGFILE
+
+   #say "script ended."  # through speaker on Mac.
+
+   trap '' EXIT INT TERM
+   exit $err 
+}
 
 command_exists() {  # in /usr/local/bin/... if installed by brew
   command -v "$@" > /dev/null 2>&1
 }
 
-### 3. Pre-requisites installation
+
+### 4. Pre-requisites installation
 
 if [[ "$INSTALL_UTILITIES" == "yes" ]]; then
 
    # Install OSX
-   # Install Homebrew
-   brew install tree
-   brew install gradle
-   brew install jq  # to handle JSON
 
-   brew install nodejs
+   if ! command_exists brew ; then
+      fancy_echo "Installing homebrew using Ruby..."
+      ruby -e "$( curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install) "
+      brew tap caskroom/cask
+   fi
+   
+   if ! command_exists tree ; then
+      brew install tree
+   fi
+   
+   if ! command_exists gradle ; then
+      brew install gradle
+   fi
+   
+   if ! command_exists jq ; then
+      brew install jq  # to handle JSON
+   fi
+   
+   if ! command_exists nodejs ; then
+      brew install nodejs
+   fi
+
    # To avoid error at 1:14 into https://app.pluralsight.com/player?course=docker-production-using-amazon-web-services&author=justin-menga&name=docker-production-using-amazon-web-services-m2&clip=3&mode=live
    node --version
    # brew install nodejs  # v9.11.1
@@ -67,15 +113,15 @@ if [[ "$INSTALL_UTILITIES" == "yes" ]]; then
    # Typesafe configuration library for JVM languages using HOCON files https://lightbend.github.io/config/
    # for friendlier 12-factor environment variable based configuration support.
 
-   ##A# Per https://github.com/wilsonmar/aws-starter
-   pip2 install ansible awscli boto3 netaddr
+   ## Per https://github.com/wilsonmar/aws-starter
+   pip install ansible awscli boto3 netaddr
 
 fi
 
-### 4. Delete local repository if it's there (for idempotency):
+### 5. Delete local repository if it's there (for idempotency):
 
 cd "$HOME"
-   echo "\n>>> Creating folder $HOME/project if it's not there:"
+   echo "\n>>> Creating folder $HOME/projects if it's not there:"
    if [ ! -d "projects" ]; then # NOT found:
       mkdir projects
    fi
@@ -85,29 +131,22 @@ cd "$HOME"
    if [ ! -d "$WORK_FOLDER" ]; then # NOT found:
       echo "\n>>> Creating folders $HOME/project/$WORK_FOLDER if it's not there:"
       mkdir "$WORK_FOLDER"
-   else
-      echo "\n>>> Removing ~/projects/$WORK_FOLDER from previous run:"
-      rm -rf "$WORK_FOLDER"
    fi
           cd "$WORK_FOLDER"
 
 echo "PWD=$PWD"
 
 
-### 5. Fork and Download dependencies
+### 6. Fork and Download dependencies
 
-WORK_REPO="microtrader"
-
-# Fork https://github.com/docker-production-aws/microtrader
-#git clone git@github.com:wilsonmar/microtrader-base.git
-#git clone git@github.com:wilsonmar/microtrader.git
-   # [1:20] https://app.pluralsight.com/player?course=docker-production-using-amazon-web-services&author=justin-menga&name=docker-production-using-amazon-web-services-m2&clip=2&mode=live
-
-      # 1:23 into https://app.pluralsight.com/player?course=docker-production-using-amazon-web-services&author=justin-menga&name=docker-production-using-amazon-web-services-m2&clip=2&mode=live
+   WORK_REPO="microtrader"
+   # 1:23 into https://app.pluralsight.com/player?course=docker-production-using-amazon-web-services&author=justin-menga&name=docker-production-using-amazon-web-services-m2&clip=2&mode=live
    echo "\n>>> Creating folder $HOME/project/$WORK_FOLDER/$WORK_REPO if it's not there:"
-   if [ ! -d "$WORK_REPO" ]; then # NOT found:
-      git clone https://github.com/docker-production-aws/microtrader.git
+   if [  -d "$WORK_REPO" ]; then # found:
+      echo "\n>>> AtRemoving ~/projects/$WORK_FOLDER/$WORK_REPO from previous run:"
+      rm -rf "$WORK_REPO"
    fi
+   git clone "https://github.com/docker-production-aws/$WORK_REPO.git"
    cd "$WORK_REPO"
    echo "PWD=$PWD"
 
@@ -117,7 +156,7 @@ git branch
 echo "PWD=$PWD"
 
 
-### 6. Build fat jars and run tests 
+### 7. Build fat jars and run tests 
 
 # 1:17 into https://app.pluralsight.com/player?course=docker-production-using-amazon-web-services&author=justin-menga&name=docker-production-using-amazon-web-services-m2&clip=3&mode=live
 ./gradlew clean test shadowJar
@@ -130,69 +169,85 @@ echo "PWD=$PWD"
 echo ">>> TODO: Stop if RETURN=$? (1 = bad/STOP, 0 = good)"
 
 
-### 7. Get rid of processes (run creates new ones):"
+### 8. Get rid of processes (run creates new ones):"
 
-echo "\n>>> Remove processes running from previous run (run creates new ones):"
+echo "\n>>> Remove disowned processes running from previous run (run creates new ones):"
+jobs  # list processes disowned.
 kill_microtrader_pids() {
    # ps -al yields: PID 32874 for CMD: /usr/bin/java -jar build/jars/microtrader-quote-2019042115340
    echo "\n>>> Killing microtrader processes running in background from previous run:"
    ps -al | grep "microtrader"
-   kill $(ps aux | grep 'microtrader-quote' | awk '{print $2}')
-   kill $(ps aux | grep 'microtrader-dashboard' | awk '{print $2}')
-   kill $(ps aux | grep 'microtrader-portfolio' | awk '{print $2}')
-   kill $(ps aux | grep 'microtrader-audit' | awk '{print $2}')
+   kill $(ps aux | grep 'microtrader-quote' | awk '{print $2}' | head -n 1 )
+   kill $(ps aux | grep 'microtrader-dashboard' | awk '{print $2}' | head -n 1 )
+   kill $(ps aux | grep 'microtrader-portfolio' | awk '{print $2}' | head -n 1 )
+   kill $(ps aux | grep 'microtrader-audit' | awk '{print $2}' | head -n 1 )
 }
 kill_microtrader_pids
 
 
-### 8. Prepare path to each Fat Jar in a different instance in & background: 
+### 9. Prepare path to each Fat Jar in a different instance in & background: 
 
+# See 1:20 into https://app.pluralsight.com/player?course=docker-production-using-amazon-web-services&author=justin-menga&name=docker-production-using-amazon-web-services-m3&clip=3&mode=live
 MAKE_VERSION=$( make version )
-echo "\n>>> MAKE_VERSION=$MAKE_VERSION"  #  MAKE_VERSION={"Version": "20190421153402.a295436"}
+   # VERBOSE: echo "\n>>> MAKE_VERSION=$MAKE_VERSION"  #  MAKE_VERSION={"Version": "20190421153402.a295436"}
 # Remove leading {"Version": " (13 characters):
 MAKE_VERSION=$( echo "${MAKE_VERSION:13}" )
 # Remove trailing "}" using POSIX command ? substitution wildcard:
 MAKE_VERSION=$( echo "${MAKE_VERSION%??}" )
-echo ">>> MAKE_VERSION=$MAKE_VERSION"  # example: 20190421153402.a295436
+echo "\n>>> MAKE_VERSION=$MAKE_VERSION"  # example: 20190421153402.a295436
 
-   QUOTE_JAR="microtrader-quote-$MAKE_VERSION-fat.jar"
-   DASHBOARD_JAR="microtrader-dashboard-$MAKE_VERSION-fat.jar"
-   PORTFOLIO_JAR="microtrader-portfolio-$MAKE_VERSION-fat.jar"
-   AUDIT_JAR="microtrader-audit-$MAKE_VERSION-fat.jar"
+# Prepare variables containing 
+   AUDIT_JAR="microtrader-audit-$MAKE_VERSION-fat.jar"          # 21,775,871
+   DASHBOARD_JAR="microtrader-dashboard-$MAKE_VERSION-fat.jar"  # 17,972,432
+   PORTFOLIO_JAR="microtrader-portfolio-$MAKE_VERSION-fat.jar"  # 17,722,025
+   QUOTE_JAR="microtrader-quote-$MAKE_VERSION-fat.jar"          # 16,723,780
 
 ls -l build/jars
       # -rw-r--r--  1 wilsonmar  staff  21775871 Jun 19 18:55 microtrader-audit-20190421153402.a295436-fat.jar
       # where "20190421153402" is Git's commitTimestamp and ".a295436" is the commitId (hash).
 
 
-### 9. Run microtrader-quote in & background: 
+### 10. Run microtrader apps in & background: 
 
+# Referenced: https://www.maketecheasier.com/run-bash-commands-background-linux/
 # 0 into https://app.pluralsight.com/player?course=docker-production-using-amazon-web-services&author=justin-menga&name=docker-production-using-amazon-web-services-m2&clip=4&mode=live
-# Start Quote Generator:
+# Start Quote Generator in background (through nohup)
 chmod +s "build/jars/$QUOTE_JAR" 
-java -jar "build/jars/$QUOTE_JAR" --cluster &
+
+java -jar "build/jars/$QUOTE_JAR" --cluster & \
+java -jar "build/jars/$DASHBOARD_JAR" --cluster & \
+java -jar "build/jars/$PORTFOLIO_JAR" --cluster & \
+java -jar "build/jars/$AUDIT_JAR" com.pluralsight.dockerproductionaws.admin.Migrate & \
+open localhost:35000 \
+fg  # fg=foreground
+
    # responses: $ Jun 20, 2019 2:13:32 AM io.vertx.core.impl.launcher.commands.RunCommand
+   # Jun 20, 2019 7:51:06 AM com.hazelcast.internal.partition.impl.PartitionStateManager
+   # INFO: [127.0.0.1]:5701 [dev] [3.10.5] Initializing cluster partition table arrangement...
+   # Jun 20, 2019 7:51:06 AM io.vertx.core.impl.launcher.commands.VertxIsolatedDeployer
+   # INFO: Succeeded in deploying verticle
+   # Market data service published : true
+   # Quotes (Rest endpoint) service published : true
+   # Server started
+   # appending output to nohup.out
 
-echo $(ps aux | grep 'microtrader-quote' | awk '{print $2}')
+echo $(ps aux | grep 'microtrader-quote' | awk '{print $2}' | head -n 1 )
 
-### Verify current stock prices appear in JSON:
+### Verify current stock prices appear in JSON: on MacOS:
 open localhost:35000
 
 exit
-
-### 10. Run microtrader-dashboard in & background: 
 
 # Start Dashboard
 java -jar "build/jars/$DASHBOARD_JAR" --cluster &
 # TODO: Open localhost:8000
 
-### 11. Run microtrader-portfolio in & background: 
 
 # Start Portfolio service to manage the stocks traded by each trader:
 java -jar "build/jars/$PORTFOLIO_JAR" --cluster &
 # View portfolio pane on dashboard localhost:8500 to verify
 
-### 12. Run microtrader-dashboard in & background after
+
 ####    creating HSQLDB schema within ./audit-db.* for Audit service:
 # 4:20 into https://app.pluralsight.com/player?course=docker-production-using-amazon-web-services&author=justin-menga&name=docker-production-using-amazon-web-services-m2&clip=4&mode=live
 java -jar "build/jars/$AUDIT_JAR" com.pluralsight.dockerproductionaws.admin.Migrate &
@@ -204,24 +259,25 @@ java -jar "build/jars/$AUDIT_JAR" com.pluralsight.dockerproductionaws.admin.Migr
 
 # Circuit breaker pattern in 7:05 into https://app.pluralsight.com/player?course=docker-production-using-amazon-web-services&author=justin-menga&name=docker-production-using-amazon-web-services-m2&clip=4&mode=live
 
+### 11. Context: Starting time stamp, OS versions, command attributes:
+
 echo "\n>>> passed"  # DEBUGGING.
 exit
 
 
 ### 13. 
 
-echo "\n>>> Run mocha tests on JavaScript:"
+echo "\n>>> Running mocha tests on JavaScript:"
 mocha --exit
-echo ">>> TODO: Stop if RETURN=$? (1 = bad/STOP, 0 = good)"
+   if [ $? -ne 0 ]; then
+      echo "/n>>> RETURN=$? (1 = bad/STOP, 0 = good)"
+   fi
+
+exit
 
 
-### 9. cd microtrader-specs  # contains package.json test
+### 14. cd microtrader-specs  # contains package.json test
 
-
-### 10. Microtrader-base
-
-# cd ..
-# git clone https://github.com/docker-production-aws/microtrader-base.git
 
 
 ### Test - see https://app.pluralsight.com/player?course=docker-production-using-amazon-web-services&author=justin-menga&name=docker-production-using-amazon-web-services-m3&clip=5&mode=live
